Uses templates for destination file names
The names of destination files often depend on the value of some
variable that is passed to a blueprint.

For example, when generating a new view template in a web application,
the name of the template file will likely include the name of the
controller and/or action for which the file will be used.