Move files are rendering blueprint
In the error case an error message was displayed twice: once for the
execution of the child command and once for the execution of the
parent command.

To mitigate this, the printing of error messages is delegated to
main.  This required changing the signature of Command.Exec to return
the actual command that has been executed.  Without returning the
actual command that has been executed, it would be impossible to tell
which command was executed by the dispatcher.